[CPDLP-4009] Moved AdditionalAdjustments::Table component to statements folder, created AdditionalAdjustments::Section component
