+++ app/components/finance/statements/additional_adjustments/section.html.erb
@@ -1,0 +1,3 @@
+<div class="finance-panel finance-panel__adjustments govuk-!-margin-top-5 govuk-!-margin-bottom-5 print-page-break">
+  <%= render Finance::Statements::AdditionalAdjustments::Table.new(statement: statement, caption_size: "m") %>
+</div>
