+++ app/components/finance/statements/additional_adjustments/table.html.erb
