+++ app/components/finance/statements/clawbacks.html.erb
@@ -1,4 +1,4 @@
-<div class="finance-panel finance-panel__adjustments govuk-!-margin-top-5 govuk-!-margin-bottom-5 print-page-break">
+<div class="finance-panel govuk-!-margin-top-5 govuk-!-margin-bottom-5">
   <%= govuk_table do |table| %>
     <% table.with_caption(size: "m", text: title) %>
